requirements-dev.txt
# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pre-commit>=3.5.0
yamlfix>=1.16.0
black>=23.12.1
flake8>=7.0.0
pyright>=1.1.0
mypy>=1.0.0
